Move dotplot code around
