+++ plugins/dotplot-view/src/DotplotRenderer/index.ts
@@ -1,3 +1,16 @@
-export { default as ReactComponent } from './components/DotplotRendering'
-export { default as configSchema } from './configSchema'
-export { default } from './DotplotRenderer'
+import PluginManager from '@jbrowse/core/PluginManager'
+import ReactComponent from './components/DotplotRendering'
+import configSchema from './configSchema'
+import DotplotRenderer from './DotplotRenderer'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addRendererType(
+    () =>
+      new DotplotRenderer({
+        name: 'DotplotRenderer',
+        configSchema: configSchema,
+        ReactComponent: ReactComponent,
+        pluginManager,
+      }),
+  )
+}
