+++ plugins/dotplot-view/src/PAFAdapter/index.ts
@@ -1,2 +1,22 @@
-export { default as AdapterClass } from './PAFAdapter'
-export { default as configSchema } from './configSchema'
+import PluginManager from '@jbrowse/core/PluginManager'
+import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
+
+import AdapterClass from './PAFAdapter'
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'PAFAdapter',
+        configSchema,
+        adapterMetadata: {
+          category: null,
+          hiddenFromGUI: true,
+          displayName: null,
+          description: null,
+        },
+        AdapterClass,
+      }),
+  )
+}
